/vsitar/: fix 94ffaf551457a660c7cdc5b04b42334911d0f46c commit (refs #1559)

+++ gdal/port/cpl_vsil_tar.cpp
@@ -374,10 +374,10 @@ int VSITarReader::GotoNextFile()
         {
             // If this is a large filename record, then read the filename
             osNextFileName.clear();
-            osNextFileName.resize(((nNextFileSize + 511) / 512) * 512);
+            osNextFileName.resize(static_cast<size_t>(((nNextFileSize + 511) / 512) * 512));
             if (VSIFReadL(&osNextFileName[0], osNextFileName.size(), 1, fp) != 1)
                 return FALSE;
-            osNextFileName.resize(nNextFileSize);
+            osNextFileName.resize(static_cast<size_t>(nNextFileSize));
             if( osNextFileName.back() == '\0' )
                 osNextFileName.resize(osNextFileName.size() - 1);
         }
